+++ libhdf5-sys/src/h5f.rs
@@ -99,6 +99,15 @@ pub enum H5F_mem_t {
     H5FD_MEM_NTYPES = 7,
 }
 
+#[cfg(not(hdf5_1_10_2))]
+#[repr(C)]
+#[derive(Copy, Clone, PartialEq, PartialOrd, Debug)]
+pub enum H5F_libver_t {
+    H5F_LIBVER_EARLIEST = 0,
+    H5F_LIBVER_LATEST = 1,
+}
+
+#[cfg(hdf5_1_10_2)]
 #[repr(C)]
 #[derive(Copy, Clone, PartialEq, PartialOrd, Debug)]
 pub enum H5F_libver_t {
@@ -109,10 +118,14 @@ pub enum H5F_libver_t {
     H5F_LIBVER_NBOUNDS = 3,
 }
 
-#[cfg(not(hdf5_1_10_0))]
-pub const H5F_LIBVER_LATEST: H5F_libver_t = H5F_LIBVER_V18;
-#[cfg(hdf5_1_10_0)]
+#[cfg(hdf5_1_10_2)]
 pub const H5F_LIBVER_LATEST: H5F_libver_t = H5F_LIBVER_V110;
+
+impl Default for H5F_libver_t {
+    fn default() -> Self {
+        H5F_LIBVER_LATEST
+    }
+}
 
 #[cfg(not(hdf5_1_10_0))]
 extern "C" {
@@ -164,6 +177,12 @@ extern "C" {
 #[cfg(hdf5_1_8_9)]
 extern "C" {
     pub fn H5Fget_file_image(file_id: hid_t, buf_ptr: *mut c_void, buf_len: size_t) -> ssize_t;
+}
+
+#[cfg(all(hdf5_1_8_9, h5_have_parallel))]
+extern "C" {
+    pub fn H5Fset_mpi_atomicity(file_id: hid_t, flag: hbool_t) -> herr_t;
+    pub fn H5Fget_mpi_atomicity(file_id: hid_t, flag: *mut hbool_t) -> herr_t;
 }
 
 #[cfg(hdf5_1_10_0)]