+++ src/globals.rs
@@ -4,7 +4,14 @@ use std::mem;
 
 use lazy_static::lazy_static;
 
-use libhdf5_sys::h5fd::{H5FD_core_init, H5FD_sec2_init, H5FD_stdio_init};
+#[cfg(h5_have_direct)]
+use libhdf5_sys::h5fd::H5FD_direct_init;
+#[cfg(h5_have_parallel)]
+use libhdf5_sys::h5fd::H5FD_mpio_init;
+use libhdf5_sys::h5fd::{
+    H5FD_core_init, H5FD_family_init, H5FD_log_init, H5FD_multi_init, H5FD_sec2_init,
+    H5FD_stdio_init,
+};
 
 use crate::internal_prelude::*;
 
@@ -315,9 +322,37 @@ lazy_static! {
 
 // File drivers
 lazy_static! {
-    pub static ref H5FD_CORE: hid_t = unsafe { H5FD_core_init() };
-    pub static ref H5FD_SEC2: hid_t = unsafe { H5FD_sec2_init() };
-    pub static ref H5FD_STDIO: hid_t = unsafe { H5FD_stdio_init() };
+    pub static ref H5FD_CORE: hid_t = unsafe { h5lock!(H5FD_core_init()) };
+    pub static ref H5FD_SEC2: hid_t = unsafe { h5lock!(H5FD_sec2_init()) };
+    pub static ref H5FD_STDIO: hid_t = unsafe { h5lock!(H5FD_stdio_init()) };
+    pub static ref H5FD_FAMILY: hid_t = unsafe { h5lock!(H5FD_family_init()) };
+    pub static ref H5FD_LOG: hid_t = unsafe { h5lock!(H5FD_log_init()) };
+    pub static ref H5FD_MULTI: hid_t = unsafe { h5lock!(H5FD_multi_init()) };
+}
+
+// MPI-IO file driver
+#[cfg(h5_have_parallel)]
+lazy_static! {
+    pub static ref H5FD_MPIO: hid_t = unsafe { h5lock!(H5FD_mpio_init()) };
+}
+#[cfg(not(h5_have_parallel))]
+lazy_static! {
+    pub static ref H5FD_MPIO: hid_t = H5I_INVALID_HID;
+}
+
+// Direct VFD
+#[cfg(h5_have_direct)]
+lazy_static! {
+    pub static ref H5FD_DIRECT: hid_t = unsafe { h5lock!(H5FD_direct_init()) };
+}
+#[cfg(not(h5_have_direct))]
+lazy_static! {
+    pub static ref H5FD_DIRECT: hid_t = H5I_INVALID_HID;
+}
+
+#[cfg(target_os = "windows")]
+lazy_static! {
+    pub static ref H5FD_WINDOWS: hid_t = *H5FD_SEC2;
 }
 
 #[cfg(test)]