+++ src/hl/plist.rs
@@ -9,6 +9,7 @@ use libhdf5_sys::h5p::{
 
 use crate::internal_prelude::*;
 
+pub mod file_access;
 pub mod file_create;
 
 /// Represents the HDF5 property list.